Add movement machanism
It's now able to move forward, backward, left and right. Also, it's
now able to turn left and right. All the inputs are triggered by
keyboard, list as below:
	W - move forward
	S - move backward
	A - move left
	D - move right
	Q - turn left
	E - turn right
It can determine if there is a wall on the moving path. However, the
begin myPosition now is fixed with [3, 3] since the maze is fixed, and
the faceDirection is all looking at 100 unit far.