Add animation of moving and turning
Add "animated" variable to switch the animation mode, "animationCount"
variable to count the current frame. However, the frame amounts of
moving and turning are both hard-coding now with 10, maybe modify in
the future.
By the way, the mechanism of setFaceDirection has been changed and
modified to a more flexible one I think.